ok now jest properly works & tests the main files
