+++ killer-sudoku/app/.babelrc
@@ -1,0 +1,5 @@
+{
+	"presets": [
+		["@nrwl/js/babel", { "useBuiltIns": "usage" }]
+	]
+}
