--- killer-sudoku/app/testing/.babelrc
@@ -1,3 +1,0 @@
-{
-	"presets": [["@nrwl/js/babel", { "useBuiltIns": "usage" }]]
-}
