+++ killer-sudoku/app/testing/babel.config.js
@@ -1,6 +1,7 @@
 module.exports = {
     presets: [
         ['@babel/preset-env', { targets: { node: 'current' } }],
+        '@babel/plugin-transform-modules-commonjs',
         '@babel/preset-react',
         '@babel/preset-flow',
     ]
