+++ killer-sudoku/app/testing/jest.config.ts
@@ -2,6 +2,10 @@ import type {Config} from 'jest';
 
 const config: Config = {
   verbose: true,
+  transform: {
+    '^.+\\\\\\\\.ts?$': 'ts-jest',
+    '^.+\\\\\\\\.tsx?$': 'ts-jest',
+  }
 };
 
 export default config;
